+++ n8n-nodes-duo/nodes/DuoSecurity/V1/DuoSecurityV1.ts
@@ -1,6 +1,6 @@
 import type { INodeTypeBaseDescription } from 'n8n-workflow';
 
-import { DuoSecurityV1Base } from './DuoSecurityV1Base.node';
+import { DuoSecurityV1Base } from './DuoSecurityV1Base';
 
 export class DuoSecurityV1 extends DuoSecurityV1Base {
 	constructor(baseDescription: INodeTypeBaseDescription) {