+++ n8n-nodes-rest-api/nodes/RestApi/V1/RestApiV1.ts
@@ -1,0 +1,9 @@
+import type { INodeTypeBaseDescription } from 'n8n-workflow';
+
+import { RestApiV1Base } from './RestApiV1Base';
+
+export class RestApiV1 extends RestApiV1Base {
+	constructor(baseDescription: INodeTypeBaseDescription) {
+		super(baseDescription);
+	}
+}